Extract Security Master mapping helpers into a partial
Increment 2g of the endpoint-monolith decomposition. Moves the Security Master DTO
mapping helpers out of WorkstationEndpoints.cs into a new sibling partial
WorkstationEndpoints.SecurityMasterMapping.cs.

Moved (5 whole private static methods, closed within the class): both MapToWorkstationSecurity
overloads (from SecuritySummaryDto and SecurityDetailDto), MapToIdentityDrillIn,
MapToEconomicDefinitionSummary, and DeriveSubType. Behavior-preserving relocation within the
single partial class. Core drops 6235 -> 6149 lines; ratchet baseline retightened.
